requirements.txt
asgiref
Django==3.2
django-allauth
djangorestframework
psycopg2-binary 
sqlparse==0.2.4
typing_extensions
mysqlclient
djongo==1.3.6
pymongo==3.12.3
django-redis
PyJWT
pytz
gunicorn